+++ app/internal/db/model_badge.go
@@ -8,15 +8,15 @@ import (
 
 // Badge - task status.
 type Badge struct {
-	ID        uint      `db:"id"`
-	ProjectID uint      `db:"project_id"`
+	ID        int64     `db:"id"`
+	ProjectID int64     `db:"project_id"`
 	Title     string    `db:"title"`
 	Config    string    `db:"complete"`
 	CreatedAt time.Time `db:"created_at"`
 }
 
 // GetID - returns the id.
-func (s *Badge) GetID() uint { return s.ID }
+func (s *Badge) GetID() int64 { return s.ID }
 
 // SetID - sets the id.
-func (s *Badge) SetID(id uint) { s.ID = id }
+func (s *Badge) SetID(id int64) { s.ID = id }
